backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
websockets==14.1
google-genai==1.14.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-cloud-secret-manager==2.21.1
google-cloud-logging==3.11.3
python-dotenv==1.0.1
pydantic==2.10.0
